requirements.txt
pytest
fastapi
faker
email-validator
python-dotenv
requests
pydantic

psycopg2-binary
sqlalchemy
sqlmodel

fastapi-pagination
uvicorn